Root Page Refactor
