+++ lib/components/navigation/nav_tab.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class NavTab{
+  const NavTab(
+      {
+      required this.label,
+      required this.icon,
+      required this.widget
+    });
+
+  final String label;
+  final Icon icon;
+  final Widget widget;
+}
