+++ lib/main.dart
@@ -2,7 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:gym_tec/auth/auth.dart';
 import 'package:gym_tec/auth/login.dart';
 import 'package:gym_tec/auth/register.dart';
-import 'package:gym_tec/homePage/user_root_page.dart';
+import 'package:gym_tec/pages/admin/admin_root_page.dart';
+import 'package:gym_tec/pages/trainer/trainer_root_page.dart';
+import 'package:gym_tec/pages/user/user_root_page.dart';
 
 void main() {
   runApp(const MyApp());
@@ -50,6 +52,8 @@ class MyApp extends StatelessWidget {
       routes: {
         '/': (context) => const AuthPage(),
         '/user': (context) => const UserRootPage(),
+        '/admin': (context) => const AdminRootPage(),
+        '/trainer': (context) => const TrainerRootPage(),
         '/login': (context) => const LoginPage(),
         '/register': (context) => const RegisterPage(),
       },
