+++ lib/pages/admin/admin_root_page.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class AdminRootPage extends StatelessWidget {
+  const AdminRootPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Text('AdminRootPage'); // Replace Text with NavSpace Widget
+  }
+}
