+++ lib/pages/trainer/trainer_root_page.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/widgets.dart';
+
+class TrainerRootPage extends StatelessWidget {
+  const TrainerRootPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Text('TrainerRootPage'); // Replace Text with NavSpace Widget
+  }
+}
