+++ lib/pages/user/routines/routine_day.dart
@@ -1,13 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:gym_tec/components/ui/padding/content_padding.dart';
 import 'package:gym_tec/components/ui/separators/context_separator.dart';
-
-import 'package:gym_tec/components/ui/separators/item_separator.dart';
-
-import 'package:gym_tec/routines/exercise_data.dart';
-import 'package:gym_tec/routines/exercise_page.dart';
-
-import '../components/ui/buttons/card_btn.dart';
+import 'package:gym_tec/pages/user/routines/exercise_data.dart';
+import 'package:gym_tec/pages/user/routines/exercise_page.dart';
+import 'package:gym_tec/components/ui/buttons/card_btn.dart';
 
 class RoutineDay extends StatelessWidget {
   RoutineDay({Key? key}) : super(key: key);
